feat: call to mind the work with exports
