+++ my-app/src/App.js
@@ -1,5 +1,10 @@
-import React from 'react';
+import React, { Component } from 'react';
+
+import { Header } from './header';
+import { Technologies } from './technologies';
+
 import './App.css';
+
 
 const App = () => {
   return (
@@ -8,29 +13,6 @@ const App = () => {
       <Technologies />
     </div >
   );
-}
-
-const Technologies = () => {
-  return (
-    <div>
-      <ul>
-        <li>css</li>
-        <li>html</li>
-        <li>js</li>
-        <li>react</li>
-      </ul>
-    </div>
-  )
-}
-
-const Header = () => {
-  return (
-    <div>
-      <a>Home</a>
-      <a>Feedback</a>
-      <a>Hello</a>
-    </div>
-  )
 }
 
 export default App;
