+++ my-app/src/header.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+
+const Header = () => {
+    return (
+      <div>
+        <a>Home</a>
+        <a>Feedback</a>
+        <a>Hello</a>
+      </div>
+    )
+  }
+
+  export { Header };
