+++ my-app/src/technologies.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+const Technologies = () => {
+    return (
+        <div>
+            <ul>
+                <li>css</li>
+                <li>html</li>
+                <li>js</li>
+                <li>react</li>
+            </ul>
+        </div>
+    )
+}
+
+
+export { Technologies };
